refactor(debug): split component gallery into /_debug/components
Theme page keeps the border/shadow scheme comparison; the component
showcase (with theme toggle) moves to its own /_debug/components route.
