+++ app/pages/_debug/theme.vue
@@ -1,11 +1,11 @@
 <template lang="pug">
 #debug-theme.body-inner
-  h1 主题调试 · 暗色硬阴影
+  h1 主题调试 · 暗色边框 / 阴影
   .toolbar
     span.toolbar__label 当前主题
     ThemeToggle
 
-  p.hint 暗色边框 / 阴影方案并排对比。每个面板自带背景、卡片、边框、阴影，互不影响，可慢慢看。点「应用到整页」会把该方案套到下方真实组件上。
+  p.hint 暗色边框 / 阴影方案并排对比。每个面板自带背景、卡片、边框、阴影，互不影响。点「应用到整页」会把该方案套到本页 CSS 变量上，可前往 #[RouterLink(to='/_debug/components') 组件展示] 查看真实组件效果。
 
   .scheme-grid
     .scheme(v-for='s in schemes' :key='s.label' :style='{ background: s.bg }')
@@ -24,79 +24,6 @@
   .live-note(v-if='applied')
     | 已把「{{ applied.label }}」实时应用到本页。
     button.scheme__apply(@click='resetLive') 还原
-
-  h2.plain.preview-title 组件预览
-  p.hint 以下组件跟随当前主题渲染；点上方某档「应用到整页预览」后，这里也会按该边框 / 阴影方案实时变化。
-
-  .preview-group
-    h3 按钮 Button
-    .row
-      FnbButton 默认
-      FnbButton(variant='primary') 主要
-      FnbButton(variant='success') 成功
-      FnbButton(variant='danger') 危险
-      FnbButton(variant='primary' :loading='true') 加载中
-      FnbButton(disabled) 禁用
-      FnbButton(variant='primary' size='sm') 小号
-      FnbButton(variant='primary' size='lg') 大号
-
-  .preview-group
-    h3 标签 Tag
-    .row
-      FnbTag 普通标签
-      FnbTag(:active='true') 选中标签
-      FnbTag(:clickable='true') 可点击
-      FnbTag(color='var(--fnb-bookmark)') 收藏
-
-  .preview-group
-    h3 卡片 Card
-    .row
-      FnbCard 白色卡片
-      FnbCard(color='brand') 品牌卡片
-      FnbCard(color='highlight') 高亮卡片
-      FnbCard(color='success') 成功卡片
-
-  .preview-group
-    h3 输入 Input / 下拉 Select
-    .row
-      FnbInput(v-model='inputVal' placeholder='请输入…')
-      FnbSelect(:options='selectOptions' v-model='selectVal')
-
-  .preview-group
-    h3 标签页 Tabs
-    FnbTabs(v-model='tabVal' :tabs='previewTabs')
-      template(#panel-tab1) 面板一内容
-      template(#panel-tab2) 面板二内容
-      template(#panel-tab3) 面板三内容
-
-  .preview-group
-    h3 消息框 Mbox
-    .stack
-      FnbMbox(type='info' header='信息') 这是一条信息提示。
-      FnbMbox(type='success' header='成功') 操作已成功完成。
-      FnbMbox(type='warning' header='警告') 请注意此处的风险。
-      FnbMbox(type='error' header='错误' :closable='true') 出现了一个错误。
-
-  .preview-group
-    h3 分页 Pagination
-    FnbPagination(:page='pageVal' :item-count='200' :page-size='20' @update:page='pageVal = $event')
-
-  .preview-group
-    h3 进度 Progress / 加载 Spin
-    .row.row--middle
-      .progress-box
-        FnbProgress(:percentage='65' :show-value='true')
-      FnbSpin(:show='true' size='medium')
-
-  .preview-group
-    h3 结果 Result
-    FnbResult(status='success' title='操作成功' description='你的请求已处理完成。')
-
-  .preview-group
-    h3 骨架屏 Skeleton
-    .row
-      FnbSkeleton
-      FnbSkeleton
 </template>
 
 <script lang="ts" setup>
@@ -116,7 +43,7 @@ const schemes: Scheme[] = [
   { label: 'A 中性白边', bg: '#2d2c31', surface: '#1f1e28', border: '#f3f3f6', shadow: '#f3f3f6' },
   { label: 'B 品牌蓝调白边', bg: '#161b26', surface: '#1d2433', border: '#f3f3f6', shadow: '#f3f3f6' },
   { label: 'C 灰双色', bg: '#14151b', surface: '#1e222b', border: '#70747f', shadow: '#3a3d47' },
-  { label: 'D 蓝硬阴影', bg: '#14151b', surface: '#1e222b', border: '#3a3d47', shadow: '#2f5ea8' },
+  { label: 'D 蓝硬阴影（当前采用）', bg: '#14151b', surface: '#1e222b', border: '#3a3d47', shadow: '#2f5ea8' },
 ]
 
 function cardStyle(s: Scheme) {
@@ -163,22 +90,6 @@ function resetLive() {
   )
   applied.value = null
 }
-
-// Component preview state
-const inputVal = ref('可编辑文本')
-const selectVal = ref('a')
-const selectOptions = [
-  { label: '选项 A', value: 'a' },
-  { label: '选项 B', value: 'b' },
-  { label: '选项 C', value: 'c' },
-]
-const tabVal = ref('tab1')
-const previewTabs = [
-  { key: 'tab1', label: '标签一' },
-  { key: 'tab2', label: '标签二' },
-  { key: 'tab3', label: '标签三' },
-]
-const pageVal = ref(2)
 
 onBeforeUnmount(resetLive)
 onMounted(() => setTitle('Debug · Theme'))
@@ -278,40 +189,5 @@ onMounted(() => setTitle('Debug · Theme'))
   align-items: center;
   gap: 1rem;
   flex-wrap: wrap;
-}
-
-.preview-title {
-  margin-top: 3.5rem;
-}
-
-.preview-group {
-  margin-top: 2rem;
-
-  h3 {
-    margin-bottom: 1rem;
-  }
-}
-
-.row {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 1rem;
-  align-items: flex-start;
-
-  &--middle {
-    align-items: center;
-  }
-}
-
-.stack {
-  display: flex;
-  flex-direction: column;
-  gap: 0.8rem;
-  max-width: 480px;
-}
-
-.progress-box {
-  flex: 1;
-  min-width: 240px;
 }
 </style>
